fix(BMDEV-2853): run execution with a New status for rejected GTC_PO first
-This is a workaround for a misplaced GTC_PO order
(like trying to place a LMT GTC_PO buy to the Ask side of the Order Book.
It is post-only so it must be placed before execution but it cannot be placed to the Ask side).
What Bitmex does in GUI looks more like rejecting with a warning message.
But what it sends via API is a common canceled order that has never been placed.
We must first give it a "Rejected" status manually and treat it like a rejected order.